Core/angular typescript skeleton (#9)
* core: add package.json file

This file contains information about the node packages used for this project. This file is
used when executing the command 'npm install'.

* core: add tsconfig file

This file is used to guide the typescript compiler as it generates JavaScript files.

* core: add typings.json file

This file is used to fetch d.ts files from external libraries which contain code/syntax that the
typescript compiler does not recognize by default. This file contains information about where to get
those d.ts files.

* core: add webpack.config.js file

This file exports the content of the config/webpack.dev.js file.

* core: add karma.conf.js file

This file export the content in the config/karma.conf.js file.

* core: add config helper file

This file helps out to export the root of the project.

* core: add common configuration file for webpack

This file contains information common to all three environments - dev, prod, and test.

* core: add webpack dev file

This file contains information specific to our dev environment.

* core: add webpack prod file

This file contains information specific to our prod environment. This file writes the bundle
to the dist folder, which we expect to to be deployed to a production server.

* core: add configuration file to run unit tests

* core: add index html file

* core: add main.ts file

This is where we bootstrap the application.

* core: add public css file

* core: add main app module

* core: add app component

* core: add app component markup and css

* test: add test file for app component

* core: add vendor file

* core: add polyfills file

* core: add .gitignore file

* move front-end sources in their own root directory