core: add linting rules
We only have a few simple rules to obey for now. These can be found in the .eslintrc file.
For example, we should indent by 4 spaces, and from looking at most of the code, we will
use single quotes. The command to run the linter (after running npm install) is npm run eslint.

Resolves #10